Proper repeatable builds (hopefully)
It will still require some configuration and tooling installation (Rust, Python, etc) but once the tooling is in place, all of the compilation of the library happens via Gradle.

No longer distributing the .so files and generated code.

Also switched away from using openssl in favor of rustls to ease compilation.