Add support for passing a client_id to convex-rs
This plumbs through the support in MobileConvexClient as well as the Kotlin ConvexClient.

The Kotlin code is now configured to read the release version from a kotlin@x.y.z Git tag. That will be used in the client_id as well as Maven packages.

The client_id will look like "kotlin-x.y.z" or "kotlin-x.y.z.dirty" for code with local changes.