Fix the Android build
It was previously unable to compile due to the newly introduced aws-lc-rs dependency in the Rust lib. It has a CMake dependency that was failing. The NDK version we were using had a bug that prevented it from choosing the right processor architecture and it also required some additional arguments passed in during the build.

This change also makes aws-lc-rs optional and when building uniffi-bindgen we leave it out to avoid compiling for yet another architecture. It's unnecessary for that phase of the build.